refactor(service): extract and enhance business logic for improved separation of concerns
