+++ src/controllers/Implementation/advisor/advisorController.ts
@@ -55,94 +55,6 @@ export default class AdvisorController implements IAdvisorController {
     }
   }
 
-  async createSlot(req: Request, res: Response): Promise<void> {
-    try {
-      if (req.body.slotData._id === '') {
-        delete req.body.slotData._id
-      }
-      const Slot = await this.advisorService.createSlot(req.body.id, req.body.slotData)
-      if (!Slot) {
-        throw new Error('Slot is already exists')
-      }
-      res.status(HttpStatusCode.CREATED).json({ message: 'Slot created successfully', Slot })
-    } catch (err) {
-      console.error(err)
-    }
-  }
-
-  async fetchSlots(req: Request, res: Response): Promise<Response> {
-    try {
-      const advisorId = req.params.advisorId
-      const page = parseInt(req.query.page as string) || 1
-      const limit = parseInt(req.query.limit as string) || 10
-      const { slots, totalPages } = await this.advisorService.fetchSlots(advisorId,page, limit)
-      return res.status(HttpStatusCode.OK).json({ success: true, data: { slots, totalPages } })
-    } catch (err) {
-      console.error(err)
-      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: "error fetching slots" })
-    }
-  }
-
-  async updateSlot(req: Request, res: Response): Promise<void> {
-    try {
-      const { slotId } = req.params;
-      const updatedSlotData = req.body;
-      const updatedSlot = await this.advisorService.updateSlot(slotId, updatedSlotData);
-      if (!updatedSlot) {
-        res.status(HttpStatusCode.NOT_FOUND).json({ message: "Slot not found" });
-      }
-      res.status(HttpStatusCode.OK).json({ message: "Slot updated successfully", slot: updatedSlot });
-    } catch (error) {
-      console.error("Error updating slot:", error);
-      res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ message: messageConstants.INTERNAL_ERROR });
-    }
-  }
-
-  async deleteSlot(req: Request, res: Response): Promise<void> {
-    try {
-      const { slotId } = req.params
-      const isDeleted = await this.advisorService.deleteSlot(slotId)
-      if (!isDeleted) {
-        res.status(HttpStatusCode.NOT_FOUND).json({ message: "cant found or delete slot" })
-      }
-      res.status(HttpStatusCode.OK).json({ message: "Slot deleted successfully" })
-    } catch (err) {
-      res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ message: messageConstants.INTERNAL_ERROR })
-    }
-  }
-
-  async getBookedSlotsForAdvisor(req: Request, res: Response): Promise<Response> {
-    try {
-      const advisorId = req.params.advisorId
-      const page = parseInt(req.query.page as string) || 1
-      const limit = parseInt(req.query.limit as string) || 10
-      const { bookedSlots, totalPages } = await this.advisorService.getBookedSlotsForAdvisor(advisorId, page, limit)
-      return res.status(HttpStatusCode.OK).json({ success: true, data: { bookedSlots, totalPages } })
-    } catch (error) {
-      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: "Error fetching booked slots" });
-    }
-  }
-
-  async fetchReviews(req: Request, res: Response): Promise<Response> {
-    try {
-      const { advisorId } = req.params;
-      const reviews = await this.advisorService.fetchReviews(advisorId);
-      return res.status(HttpStatusCode.OK).json({ success: true, data: reviews });
-    } catch (error) {
-      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: 'Error fetching reviews' })
-    }
-  }
-
-  async addReplyToReview(req: Request, res: Response): Promise<Response> {
-    try {
-      const { reviewId } = req.params
-      const { advisorId, text } = req.body
-      const review = await this.advisorService.addReplyToReview(reviewId, advisorId, text)
-      return res.status(HttpStatusCode.OK).json({ success: true, data: review })
-    } catch (err) {
-      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: 'Error adding review' })
-    }
-  }
 
   async fetchDashboard(req: Request, res: Response): Promise<Response> {
     try {
@@ -214,5 +126,33 @@ export default class AdvisorController implements IAdvisorController {
       return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ message: messageConstants.INTERNAL_ERROR });
     }
   }
+
+  async fetchAdvisors(req: Request, res: Response): Promise<Response> {
+    try {
+      const page = parseInt(req.query.page as string) || 1;
+      const limit = parseInt(req.query.limit as string) || 10;
+      const { users, totalPages } = await this.advisorService.fetchAdvisors(page, limit);
+      return res.status(HttpStatusCode.OK).json({ success: true, data: { users, totalPages } });
+    } catch (error) {
+      console.error(error)
+      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: messageConstants.INTERNAL_ERROR    });
+    }
+  }
+
+  async updateAdvisorBlockStatus(req:Request, res:Response):Promise<Response>{
+    try{
+      const {action,email} = req.body
+      const result = await this.advisorService.updateAdvisorBlockStatus(action,email)
+      if(result.error){
+        return res.status(HttpStatusCode.BAD_REQUEST).json({error: result.error})
+      }
+      return res.status(HttpStatusCode.OK).json({message:result.message})
+    }catch(error){
+      console.error(error)
+      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({error: messageConstants.INTERNAL_ERROR   })
+    }
+  }
+
+
 
 }
