+++ src/controllers/Implementation/category/categoryController.ts
@@ -3,6 +3,7 @@ import { inject, injectable } from "tsyringe";
 import { ICategoryController } from "../../Interface/category/ICategoryController";
 import { HttpStatusCode } from '../../../utils/httpStatusCode';
 import { ICategoryService } from '../../../services/Interface/category/ICategoryService';
+import { messageConstants } from '../../../utils/messageConstants';
 
 @injectable()
 export default class CategoryController implements ICategoryController {
@@ -18,6 +19,58 @@ export default class CategoryController implements ICategoryController {
     } catch (err) {
       console.error(err)
       return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ messag: 'Error fetching categories' })
+    }
+  }
+
+  async fetchCategories(req: Request, res: Response): Promise<Response> {
+    try {
+      const page = parseInt(req.query.page as string) || 1;
+      const limit = parseInt(req.query.limit as string) || 10;
+      const { categories, totalPages } = await this.categoryService.fetchCategories(page, limit);
+      return res.status(HttpStatusCode.OK).json({ success: true, data: { categories, totalPages } });
+    } catch (error) {
+      console.error(error)
+      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ error: messageConstants.INTERNAL_ERROR    });
+    }
+  }
+
+  async addCategory(req: Request, res: Response): Promise<Response> {
+    try {
+        const { name } = req.body;
+        const category = await this.categoryService.addCategory(name);
+        return res.status(HttpStatusCode.CREATED).json({ success: true, data: { category } });
+    } catch (err: any) {
+        console.error(err);
+        if (err.message === "CATEGORY_EXISTS") {
+            return res.status(HttpStatusCode.BAD_REQUEST).json({ success: false, error: "Category already exists" });
+        }
+        return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({ success: false, error: messageConstants.INTERNAL_ERROR    });
+    }
+}
+
+  async updateCategory(req:Request,res:Response):Promise<Response>{
+    try{
+      const {id} = req.params
+      const {name} = req.body
+      const updatedCategory = await this.categoryService.updateCategory(id,name)
+      return res.status(HttpStatusCode.OK).json({success:true,data:{updatedCategory}})
+    }catch(err){
+      console.error(err)
+      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({error: messageConstants.INTERNAL_ERROR   })
+    }
+  }
+
+  async deleteCategory(req:Request, res:Response):Promise<Response>{
+    try{
+      const {id} = req.params
+      const category = await this.categoryService.deleteCategory(id)
+      if(!category){
+        return res.status(HttpStatusCode.NOT_FOUND).json({message:"category not found"})
+      }
+      return res.status(HttpStatusCode.OK).json({message:"category deleted successfully"})
+    }catch(err){
+      console.error(err)
+      return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({error:"Error deleting category"})
     }
   }
 }
