+++ src/controllers/Interface/complaint/IComplaintController.ts
@@ -2,5 +2,5 @@ import {Request,Response} from 'express'
 
 export interface IComplaintController{
     reportAdvisor(req:Request, res:Response): Promise<Response> 
-
+    fetchReports(req:Request, res:Response): Promise<Response>
 }
