+++ src/repositories/Implementation/slotRepository.ts
@@ -15,7 +15,7 @@ export default class SlotRepository extends BaseRepository<Slot>implements ISlot
         const bookedSlot = await slotSchema.findOneAndUpdate({ _id: slotId }, slot, { new: true })
         return bookedSlot
     }
-    async updateSlot(slotId: string): Promise<Slot | null> {
+    async updateSlotStatus(slotId: string): Promise<Slot | null> {
         return await slotSchema.findOneAndUpdate({ _id: slotId }, { status: "Cancelled" }, { new: true })
     }
     async fetchSlotsByUser(userId: string, page: number, limit: number): Promise<{ slots: Slot[], totalPages: number }> {
@@ -34,6 +34,58 @@ export default class SlotRepository extends BaseRepository<Slot>implements ISlot
             return { slots, totalPages };
         } catch (error: any) {
             throw new Error(`Error fetching slots: ${error.message}`);
+        }
+    }
+
+    async createSlot(slot: Slot): Promise<Slot> {
+        const result = await slotSchema.create(slot)
+        return result
+    }
+
+    async findExistingSlot(date: string, startTime: string): Promise<boolean> {
+        const result = await slotSchema.findOne({ date, startTime })
+        return !!result
+    }
+
+    async fetchSlots(advisorId:string,page: number, limit: number): Promise<{ slots: Slot[] | Slot; totalSlots: number }> {
+        const skip = (page - 1) * limit
+        const [slots, totalSlots] = await Promise.all([
+            slotSchema.find({ "advisorId._id": advisorId }).skip(skip).limit(limit),
+            slotSchema.countDocuments()
+        ])
+        return { slots, totalSlots }
+    }
+
+    async findSlotById(slotId: string): Promise<Slot | null> {
+        return await slotSchema.findById(slotId)
+    }
+
+    async updateSlot(slotId: string, slot: Slot): Promise<Slot | null> {
+        return await slotSchema.findByIdAndUpdate(slotId, slot, { new: true })
+    }
+
+    async deleteSlot(slotId: string): Promise<boolean> {
+        const result = await slotSchema.findByIdAndDelete(slotId);
+        return !!result;
+    }
+
+    async getBookedSlotsForAdvisor(advisorId: string, page: number, limit: number): Promise<{ bookedSlots: Slot[] | Slot; totalSlots: number }> {
+        // eslint-disable-next-line no-useless-catch
+        try {
+            const skip = (page - 1) * limit
+            const [bookedSlots, totalSlots] = await Promise.all([
+                slotSchema
+                    .find({ 'advisorId._id': advisorId, status: "Booked" })
+                    .skip(skip)
+                    .limit(limit)
+                    .populate("bookedBy", "username email")
+                    .exec(),
+
+                slotSchema.countDocuments({ 'advisorId._id': advisorId, status: "Booked" })
+            ]);
+            return { bookedSlots, totalSlots };
+        } catch (err) {
+            throw err
         }
     }
 }
