+++ src/repositories/Interface/IComplaintRepository.ts
@@ -2,5 +2,6 @@ import { IReport } from "../../models/reportSchema";
 
 export interface IComplaintRepository{
     createReport(data:IReport):Promise<IReport>
+    fetchReports(page:number,limit:number):Promise<{reports:IReport[], totalReports:number}>
 
 }
